requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
datasets==2.16.1
huggingface-hub==0.20.3
